+++ client/cmd/remote-docker/machine.go
@@ -18,9 +18,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/lhns/remote-docker/client/internal/config"
-	"github.com/lhns/remote-docker/client/internal/machine"
 	"github.com/lhns/remote-docker/client/internal/proxy"
 	"github.com/lhns/remote-docker/core-client/keys"
+	"github.com/lhns/remote-docker/machine"
 )
 
 func newMachineCommand() *cobra.Command {